Probably the Final Commit of PizzaPedidoForm
This project was incomplete because the price of the pizza selected by the costumer was not being showed at the 'pedido.html' page. At first, I tried to export the colossal function I created when I was quite unexperienced at programming (The function named calcularPreco() ). Anyway, to calculate the price in order to show it to the client at the 'pedido.html' page, I created another function, which is still a long one, but quite smaller and simpler than the 'calcularPreco()'. That's pretty much it! :)